UI proj mat & use f32s for text rendering

+++ engine/font.rs
@@ -2,17 +2,17 @@ use std::{fs, io::{self, Read, Seek}};
 
 pub struct Glyph {
 	pub char_code: u32,
-	pub position: (u32, u32),
-	pub size: (u32, u32),
-	pub bearing: (i32, i32),
-	pub advance: i32
+	pub position: (f32, f32),
+	pub size: (f32, f32),
+	pub bearing: (f32, f32),
+	pub advance: f32
 }
 
 pub struct Font {
 	pub file_path: String,
 	pub atlas_width: u32,
 	pub atlas_height: u32,
-	pub space_advance: i32,
+	pub space_advance: f32,
 	pub glyphs: Vec<Glyph>
 }
 
@@ -30,7 +30,7 @@ impl Font {
 		file.seek(io::SeekFrom::Current((atlas_width * atlas_height) as i64)).unwrap();
 
 		file.read_exact(&mut bytes).unwrap();
-		let space_advance = i32::from_ne_bytes(bytes);
+		let space_advance = f32::from_ne_bytes(bytes);
 
 		file.read_exact(&mut bytes).unwrap();
 		let glyph_count = u32::from_ne_bytes(bytes) as usize;
@@ -47,25 +47,25 @@ impl Font {
 			let char_code = u32::from_ne_bytes(bytes);
 
 			bytes.copy_from_slice(&buffer[glyph_offset + 4..glyph_offset + 8]);
-			let position_x = u32::from_ne_bytes(bytes);
+			let position_x = f32::from_ne_bytes(bytes);
 
 			bytes.copy_from_slice(&buffer[glyph_offset + 8..glyph_offset + 12]);
-			let position_y = u32::from_ne_bytes(bytes);
+			let position_y = f32::from_ne_bytes(bytes);
 
 			bytes.copy_from_slice(&buffer[glyph_offset + 12..glyph_offset + 16]);
-			let width = u32::from_ne_bytes(bytes);
+			let width = f32::from_ne_bytes(bytes);
 
 			bytes.copy_from_slice(&buffer[glyph_offset + 16..glyph_offset + 20]);
-			let height = u32::from_ne_bytes(bytes);
+			let height = f32::from_ne_bytes(bytes);
 
 			bytes.copy_from_slice(&buffer[glyph_offset + 20..glyph_offset + 24]);
-			let bearing_x = i32::from_ne_bytes(bytes);
+			let bearing_x = f32::from_ne_bytes(bytes);
 
 			bytes.copy_from_slice(&buffer[glyph_offset + 24..glyph_offset + 28]);
-			let bearing_y = i32::from_ne_bytes(bytes);
+			let bearing_y = f32::from_ne_bytes(bytes);
 
 			bytes.copy_from_slice(&buffer[glyph_offset + 28..glyph_offset + 32]);
-			let advance = i32::from_ne_bytes(bytes);
+			let advance = f32::from_ne_bytes(bytes);
 
 			glyphs.push(Glyph {
 				char_code,
